+++ CHANGES.md
@@ -1,6 +1,12 @@
 Changes
 =======
 
+2013/05/03 pfraze
+
+ - Added inline style sanitization/whitelisting
+   - supports padding/padding-* (clamped), margin/margin-* (clamped), color, background, font, font-size, font-style, font-weight, line-height, letter-spacing, text-align, text-decoration, border/border-*, box-shadow, vertical-align, overflow
+
+
 2013/05/02 pfraze
  
  - Added request link header with user 'storage' href